create shared.py where logger is
